requirements.txt
numpy
imutils
opencv-python
Pillow
carvekit
flask
